+++ include/jsonparser.h
@@ -28,7 +28,7 @@ typedef enum {
 	RT_ALLOC_STRING,		// we expect a string in the JSON input and will return a malloced version of the string in the target object 
 	RT_CHAR,				// we expect a number in the JSON input, which will define a character value
 	RT_CHAR_ZERO,			// we expect a number in the JSON input, which will define a character value. 0 characters are explicitly saved.
-	RT_FLAG,				// we expect a boolean value in the JSON input. If true, we will use the 
+	RT_FLAG,				// we expect a boolean value in the JSON input. If true, we will use the rt_flag field as a bit value to or with the int value.
 	RT_INTEGER,				// we expect an "int" value in the JSON input.
 	RT_FLOAT,				// we expect a "float" value in the JSON input.
 	RT_INTEGER_ARRAY,		// we expect an array of int values in the JSON input and will assume the target object to contain an int[] data structure to fill.